comment

+++ database/migrations/2023_11_13_001913_create_shippings_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
 
             $table->string('area');
-            $table->decimal('price',15,2);
+            $table->float('price',15,2);
             $table->enum('status',['active','inactive'])->default('active');
 
             $table->timestamps();
